# docs/swagger/polaris-api.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "",
        "version": "8.0",
        "title": "Polaris",
        "termsOfService": ""
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
...
    "paths": {
...
        "/trigger_index": {
            "post": {
                "tags": [
                    "Other"
                ],
                "summary": "Begins or queues a crawl of the music collection",
                "operationId": "postTriggerIndex",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/config": {
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Amends the server settings, mount directories and list of users",
                "operationId": "putConfig",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Config"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/mount_dirs": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads the existing collection mount directories",
                "operationId": "getMountDirs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#components/schemas/MountDir"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Replaces the list collection mount directories",
                "operationId": "putMountDirs",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "$ref": "#components/schemas/MountDir"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/settings": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads the existing server settings",
                "operationId": "getSettings",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/Settings"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Amends the server settings",
                "operationId": "putSettings",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Settings"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List existing user accounts",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Creates a new user account",
                "operationId": "postUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/NewUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/user/{name}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates properties of an existing user",
                "operationId": "putUserName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the affected user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/UserUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletes an existing user",
                "operationId": "deleteUserName",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of the affected user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "admin_http_bearer": [],
                        "admin_query_parameter": []
                    }
                ]
            }
        },
        "/preferences": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Reads the preferences of the current user",
                "operationId": "getPreferences",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/Preferences"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_http_bearer": [],
                        "auth_query_parameter": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Saves the preferences of the current user",
                "operationId": "putPreferences",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Preferences"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "auth_http_bearer": [],
                        "auth_query_parameter": []
                    }
                ]
            }
        },
        "/auth": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Signs in a user.",
                "operationId": "postAuth",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/Credentials"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Authorization"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/browse": {
            "get": {
                "tags": [
                    "Collection"
                ],
                "summary": "Reads the content of the top-level directory in the music collection",
                "operationId": "getBrowse",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/APIVersion"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CollectionFile"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_http_bearer": [],
                        "auth_query_parameter": []
                    }
                ]
            }
        },
...
        "/playlists": {
            "get": {
                "tags": [
                    "Playlists"
                ],
                "summary": "Lists the playlists belonging to the current user",
                "operationId": "getPlaylists",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ListPlaylistsEntry"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth_http_bearer": [],
                        "auth_query_parameter": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "is_admin": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "name",
                    "is_admin"
                ]
            },
...
        },
...
        "securitySchemes": {
            "auth_http_bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "An authentication token obtained in the output of the `auth` endpoint"
            },
            "admin_http_bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Identical to the auth_http_bearer scheme but only for users recognized as admin by the Polaris server"
            },
            "auth_query_parameter": {
                "type": "apikey",
                "in": "query",
                "name": "auth_token",
                "description": "An authentication token obtained in the output of the `auth` endpoint"
            },
            "admin_query_parameter": {
                "type": "apikey",
                "in": "query",
                "name": "auth_token",
                "description": "Identical to the auth_query_parameter scheme but only for users recognized as admin by the Polaris server"
            }
        },
...
    },
    "security": []
}